Changed slf4j to 2.0.12 as that supports modules, updates modules-info with exports and requires

+++ easy-rules-core/src/test/java/org/jeasy/rules/api/FactsTest.java
@@ -25,6 +25,7 @@ package org.jeasy.rules.api;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -39,7 +40,7 @@ public class FactsTest {
         facts.put("foo", 2);
 
         assertThat(facts).hasSize(1);
-        assertThat(facts.get("foo")).isEqualTo(2);
+        assertThat((BigDecimal) facts.get("foo")).isEqualTo(2);
     }
 
     @Test
@@ -72,7 +73,7 @@ public class FactsTest {
     @Test
     public void get() throws Exception {
         facts.put("foo", 1);
-        assertThat(facts.get("foo")).isEqualTo(1);
+        assertThat((BigDecimal) facts.get("foo")).isEqualTo(1);
     }
 
     @Test
